Fixed any branch issues
